requirements.txt
fastapi
uvicorn
pydantic
python-multipart
torch
transformers[torch]
scikit-learn
pandas
numpy
biopython
shap
joblib